# frontend/src/swagger-api/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEmailLoginDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthEmailLoginResponseDto"}}}}},"tags":["Auth"]}},"/api/v1/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"]},"patch":{"operationId":"AuthController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthUpdateDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshResponseDto"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/"}}}},"204":{"description":""}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/api/v1/users/create-employee":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"Create a new employee","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDataSwagger"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/users":{"get":{"operationId":"UsersController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"filters","required":false,"in":"query","schema":{"nullable":true,"type":"string"}},{"name":"sort","required":false,"in":"query","schema":{"nullable":true,"type":"string"}},{"name":"searchText","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Get all users","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InfinityPaginationResultTypeSwagger"}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/users/employees":{"get":{"operationId":"UsersController_findAllEmployees","parameters":[],"responses":{"200":{"description":"Get all employees","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/v1/users/{id}":{"patch":{"operationId":"UsersController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["Users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"Delete an employee"},"204":{"description":""}},"tags":["Users"],"security":[{"bearer":[]}]}},"/api/tasks":{"post":{"operationId":"TasksController_create","summary":"Create a new task","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaskDto"}}}},"responses":{"201":{"description":"Task created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDataSwagger"}}}},"403":{"description":"Only employers can create tasks"}},"tags":["Tasks"],"security":[{"bearer":[]}]},"get":{"operationId":"TasksController_findAll","summary":"Get all tasks with pagination and filtering","parameters":[{"name":"assigneeId","required":false,"in":"query","description":"Assignee user ID filter","schema":{"type":"number"}},{"name":"statusId","required":false,"in":"query","description":"Status ID filter","schema":{"type":"number"}},{"name":"sortBy","required":false,"in":"query","description":"Sort by field","schema":{"enum":["createdAt","dueDate","status"],"type":"string"}},{"name":"sortOrder","required":false,"in":"query","description":"Sort order","schema":{"enum":["ASC","DESC"],"type":"string"}},{"name":"page","required":false,"in":"query","description":"Page number","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","description":"Page size","schema":{"type":"number"}}],"responses":{"200":{"description":"Tasks retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskPaginationResultSwagger"}}}}},"tags":["Tasks"],"security":[{"bearer":[]}]}},"/api/tasks/summary":{"get":{"operationId":"TasksController_getTaskSummary","summary":"Get task summary for all employees (paginated)","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}}],"responses":{"200":{"description":"Task summary retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaginateEmployeeTaskSummaryDto"}}}},"403":{"description":"Only employers can view task summaries"}},"tags":["Tasks"],"security":[{"bearer":[]}]}},"/api/tasks/{id}":{"get":{"operationId":"TasksController_findOne","summary":"Get a task by ID","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Task retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDataSwagger"}}}},"403":{"description":"Access denied"},"404":{"description":"Task not found"}},"tags":["Tasks"],"security":[{"bearer":[]}]},"patch":{"operationId":"TasksController_update","summary":"Update a task","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaskDto"}}}},"responses":{"200":{"description":"Task updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDataSwagger"}}}},"403":{"description":"Access denied"},"404":{"description":"Task not found"}},"tags":["Tasks"],"security":[{"bearer":[]}]},"delete":{"operationId":"TasksController_remove","summary":"Delete a task","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Task deleted successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponseDataSwagger"}}}},"403":{"description":"Only employers can delete tasks"},"404":{"description":"Task not found"}},"tags":["Tasks"],"security":[{"bearer":[]}]}}},"info":{"title":"API","description":"API docs","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"Role":{"type":"object","properties":{"id":{"type":"number","example":1},"name":{"type":"string","example":"Admin"}},"required":["id"]},"User":{"type":"object","properties":{"email":{"type":"string","nullable":true},"name":{"type":"string","nullable":true},"role":{"$ref":"#/components/schemas/Role"},"roleId":{"type":"number"},"password":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"deletedAt":{"format":"date-time","type":"string"},"id":{"type":"number"}},"required":["email","name","role","roleId","createdAt","updatedAt","deletedAt","id"]},"AuthEmailLoginResponseDto":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"tokenExpires":{"type":"string"},"user":{"$ref":"#/components/schemas/User"},"infoTokenStore":{"type":"object"}},"required":["token","refreshToken","tokenExpires","user","infoTokenStore"]},"AuthEmailLoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"RefreshResponseDto":{"type":"object","properties":{"token":{"type":"string"},"refreshToken":{"type":"string"},"tokenExpires":{"type":"number"}},"required":["token","refreshToken","tokenExpires"]},"AuthUpdateDto":{"type":"object","properties":{"firstName":{"type":"string","example":"John"},"lastName":{"type":"string","example":"Doe"},"email":{"type":"string","example":"new.email@example.com"},"password":{"type":"string","minLength":6},"oldPassword":{"type":"string","minLength":6}}},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"test1@example.com"},"password":{"type":"string","minLength":6,"example":"password123"}},"required":["name","email","password"]},"ResponseDataSwagger":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"},"data":{"type":"object"}},"required":["statusCode","message","data"]},"InfinityPaginationResultTypeSwagger":{"type":"object","properties":{"total":{"type":"number"},"data":{"type":"array","items":{"type":"array"}}},"required":["total","data"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"test1@example.com"}}},"CreateTaskDto":{"type":"object","properties":{"title":{"type":"string","description":"Task title"},"description":{"type":"string","description":"Task description"},"assigneeId":{"type":"number","description":"Assignee user ID"},"statusId":{"type":"number","description":"Status ID"},"dueDate":{"type":"string","description":"Due date"}},"required":["title","assigneeId"]},"Status":{"type":"object","properties":{"id":{"type":"number","example":1},"name":{"type":"string","example":"Active"}},"required":["id"]},"Task":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"description":{"type":"string","nullable":true},"status":{"$ref":"#/components/schemas/Status"},"assignee":{"$ref":"#/components/schemas/User"},"createdBy":{"$ref":"#/components/schemas/User"},"createdAt":{"format":"date-time","type":"string"},"dueDate":{"format":"date-time","type":"string","nullable":true},"assigneeId":{"type":"number"},"statusId":{"type":"number"},"deletedAt":{"format":"date-time","type":"string"}},"required":["id","title","description","status","assignee","createdBy","createdAt","dueDate","deletedAt"]},"TaskPaginationResultSwagger":{"type":"object","properties":{"total":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/Task"}}},"required":["total","data"]},"TaskSummary":{"type":"object","properties":{"userId":{"type":"number"},"userName":{"type":"string"},"totalTasks":{"type":"number"},"completedTasks":{"type":"number"}},"required":["userId","userName","totalTasks","completedTasks"]},"PaginateEmployeeTaskSummaryDto":{"type":"object","properties":{"total":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TaskSummary"}}},"required":["total","data"]},"UpdateTaskDto":{"type":"object","properties":{"title":{"type":"string","description":"Task title"},"description":{"type":"string","description":"Task description"},"assigneeId":{"type":"number","description":"Assignee user ID"},"statusId":{"type":"number","description":"Status ID"},"dueDate":{"type":"string","description":"Due date"}}}}}}